+++ src/lib/converter/plugins/SourcePlugin.ts
@@ -10,6 +10,7 @@ import { getCommonDirectory, normalizePath } from "../../utils/fs";
 import { dirname, relative } from "path";
 import { SourceReference } from "../../models";
 import { gitIsInstalled, Repository } from "../utils/repository";
+import { BasePath } from "../utils/base-path";
 
 /**
  * A handler that attaches source file information to reflections.
@@ -77,7 +78,7 @@ export class SourcePlugin extends ConverterComponent {
         const symbol = reflection.project.getSymbolFromReflection(reflection);
         for (const node of symbol?.declarations || []) {
             const sourceFile = node.getSourceFile();
-            const fileName = sourceFile.fileName;
+            const fileName = BasePath.normalize(sourceFile.fileName);
             this.fileNames.add(fileName);
 
             let position: ts.LineAndCharacter;
@@ -115,7 +116,7 @@ export class SourcePlugin extends ConverterComponent {
         if (this.disableSources || !sig) return;
 
         const sourceFile = sig.getSourceFile();
-        const fileName = sourceFile.fileName;
+        const fileName = BasePath.normalize(sourceFile.fileName);
         this.fileNames.add(fileName);
 
         const position = ts.getLineAndCharacterOfPosition(