+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+### Bug Fixes
+
+-   Normalize all file paths on Windows, #2113.
+
 ## v0.23.21 (2022-11-14)
 
 ### Features